Migrate all financial calculations from double/float to BigDecimal
XOF currency amounts use zero decimal places, percentages and index
values use two decimal places with RoundingMode.HALF_UP. This ensures
precise financial arithmetic across Quote, OrderBookEntry, MarketIndex,
and MarketDataSimulator.

+++ backend/src/main/java/com/bloomfield/terminal/model/MarketIndex.java
@@ -1,11 +1,12 @@
 package com.bloomfield.terminal.model;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record MarketIndex(
         String name,
-        double value,
-        double change,
-        double changePercent,
-        List<Double> sparklineData
+        BigDecimal value,
+        BigDecimal change,
+        BigDecimal changePercent,
+        List<BigDecimal> sparklineData
 ) {}
